refactor: 예외 클래스 수정
예외 클래스 수정
- 예외 패키지 구조 일부 수정

related to: #276
